refactor: drinks/indexのロジックをstoreに移行した
